+++ PokemonProject/src/pokemons/SmallPokemon.java
@@ -1,7 +1,5 @@
 package pokemons;
 
-import attacks.PokemonAttack;
-
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -10,7 +8,7 @@ public class SmallPokemon extends Pokemon {
     public SmallPokemon(String name, List<String> types ) {
         this.name = name;
         this.types = types;
-        this.hp = this.returnInitialHP(); // can do it because defaultHp is package private
+        this.hp = this.returnInitialHP();
         this.attackPoints = this.returnInitialAttackPoints();
         this.defencePoints = this.returnInitialDefencePoints();
     }
@@ -24,19 +22,19 @@ public class SmallPokemon extends Pokemon {
         this.defencePoints = this.returnInitialDefencePoints();
     }
 
-
+// methods
     @Override
-    public double returnInitialHP() {
-        return 0.8 * Pokemon.defaultHp;
+    public int returnInitialHP() {
+        return (int) (0.8 * Pokemon.defaultHp);
     }
 
     @Override
-    public double returnInitialAttackPoints() {
-        return 0.8 * Pokemon.defaultAttackPoints;
+    public int returnInitialAttackPoints() {
+        return (int) (0.8 * Pokemon.defaultAttackPoints);
     }
 
     @Override
-    public double returnInitialDefencePoints() {
-        return 0.8 * Pokemon.defaultDefencePoints;
+    public int returnInitialDefencePoints() {
+        return (int) (0.8 * Pokemon.defaultDefencePoints);
     }
 }
